requirements.txt
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-Cors==5.0.0
Flask-MySQL==1.6.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
gspread==6.1.2
gspread-formatting==1.2.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==9.1.0
oauthlib==3.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyMySQL==1.1.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
urllib3==2.2.3
Werkzeug==3.0.4
